== Combined Other small changes I made on.
-- 1.TAKING YOUR FILE INTO NEW PAGE AND THEN RENAME
=====================================================
-- I convert the name of page2.dart & take into new file called into
import_card and it is new screen that pops up when we click the
addCards button.

-- I also changed your code  how?
-- previous it was row of three containers with thier design and
-- child of each container has seperate Icon (camera icon, gallery icon
-- and search icon).
-- since all of them have same desing with only the icon and changed into
what is call builder() - in user defined functions that returns widget
and takes argument of dynamic value like now I am passing icon then
it return the same design with any icon you pass.

-- you can look the code.

-- 2. ADDED NEW FONT FOR TITLE.
================================================================
 I add 'Roboto" font for title and make the defualt font of app
 using the Theme in the main.dart